requirements.txt
#install a virtualenv
# load it using source venv/bin/activate
# install the requirements using pip install -r requirements.txt
openai
python-dotenv
langchain
langchain-openai
jupyterlab
beautifulsoup4
faiss-cpu # vector store